Prepare for transition to v3:  * Module dependency calculation now done in `forge.dependency`  * Created `forge.views` namespace  * Moved v1 specific views to `forge.views.v1`  * Moved JSON-specific views to `forge.views.json`
