+++ forge/views/v1.py
@@ -1,36 +1,8 @@
-from django.http import HttpResponse
 from django.shortcuts import render
-from django.utils import simplejson
 
-from .models import Module, Release
-
-
-def json_response(data, indent=None, status=None):
-    return HttpResponse(simplejson.dumps(data, indent=indent),
-                        mimetype='application/json',
-                        status=status)
-
-
-def error_response(errors, **kwargs):
-    kwargs.setdefault('status', 400)
-    # The Puppet Forge sometimes returns it's errors in one of two ways,
-    # as a list:
-    #   {"errors": [ "<msg1>" ... "<msgN>" ]}
-    # Or as a single element:
-    #   {"error": "<msg>"}
-    if isinstance(errors, (list, tuple)):
-        key = 'errors'
-    else:
-        key = 'error'
-    return json_response({key: errors}, **kwargs)
-
-
-def handler404(request):
-    return render(request, 'admin/404.html', {})
-
-
-def handler500(request):
-    return render(request, 'admin/500.html', {})
+from .json import error_response, json_response
+from ..dependency import module_dependencies
+from ..models import Module, Release
 
 
 def module_dict(module):
@@ -121,50 +93,14 @@ def releases_json(request):
     version = request.GET.get('version', None)
     if version:
         try:
-            mod_release = Release.objects.get(module=module, version=version)
+            release = Release.objects.get(module=module, version=version)
         except Release.DoesNotExist:
             return error_response('Module %s has no release for version %s' %
                                   (full_name, version), status=410)
     else:
-        mod_release = None
+        release = None
 
-    # This is the code that figures out dependencies; there is much
-    # room for improvement.
-    dependencies = {}
-    queue = set()
-
-    def populate_dependencies(mod, release=None):
-        if release is None:
-            release = mod.latest_release
-
-        release_dependencies = [
-            (depend['name'], depend.get('version_requirement', '>= 0.0.0'))
-            for depend in release.metadata.get('dependencies', [])
-        ]
-        full_name = str(release.module)
-        release_info = {
-            'version': str(release.version),
-            'file': release.tarball.url,
-            'dependencies': release_dependencies,
-        }
-
-        if dependencies.get(full_name, None):
-            dependencies[full_name].append(release_info)
-        else:
-            dependencies[full_name] = [release_info]
-
-        for name, version in release_dependencies:
-            if not name in dependencies:
-                queue.add(name)
-
-    # Keep going until we don't have any dependencies left.
-    populate_dependencies(module, mod_release)
-    while queue:
-        full_name = queue.pop()
-        try:
-            populate_dependencies(Module.get_for_full_name(full_name))
-        except Module.DoesNotExist:
-            err = 'Dependency module %s not found' % full_name
-            return error_response([err], status=410)
-
-    return json_response(dependencies)
+    try:
+        return json_response(module_dependencies(module, release))
+    except Exception as e:
+        return error_response([e.message], status=410)
